requirements-local-voice.txt
litert-lm-api
faster-whisper>=1.0.0,<2.0.0
pyaudio>=0.2.14
webrtcvad>=2.0.10
python-dotenv>=1.0.0
pandas>=2.0.0,<3.0.0
numpy>=1.24.0,<2.0.0
pyyaml>=6.0
coloredlogs>=15.0
